Added to have parameters.xml

+++ JAVA/Main.java
@@ -5,85 +5,69 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.io.File;
 import java.io.IOException;
+import javax.xml.parsers.DocumentBuilder;
+import javax.xml.parsers.DocumentBuilderFactory;
+import org.w3c.dom.Document;
+import org.w3c.dom.Element;
 
 public class Main {
 
     private static double[] function(double[] X) {
         double x = X[0];
-        if(x > 50){
+        if (x > 50) {
             x = x - 50;
         }
         x = Math.sin(x);
-        return new double[] { x };
+        return new double[]{x};
     }
 
     public static void main(String[] args) throws IllegalAccessException, IOException {
+        if (args.length != 1) {
+            System.out.println("Usage: java JAVA.Main <path to parameters.xml>");
+            return;
+        }
+
+        String xmlFilePath = args[0];
+        Parameters params = readParametersFromXML(xmlFilePath);
 
         Other.deleteAll(new File("PYTHON/images"));
 
-        // Initializer.setSeed(42);
-
-        int feature = 1;
-        int output = 1;
-
-        boolean video = !true && feature == 1 && output == 1;
-
-        int train = 3000;
-        int test = train / 5;
-        int validate = train / 10;
-
-        double lower_bound = -2, upper_bound = 20;
+        boolean video = true && params.feature == 1 && params.output == 1;
 
         Random rand = new Random();
 
-        Integer[] structure = { feature, 30, 30, 30, 30, output };
-        String[] functions = { "linear", "reLU", "reLU", "reLU","reLU","linear" };
+        Model model = new Model(params.type, params.structure, params.functions, params.epochs, params.lr, params.Model,
+                params.beta1, params.beta2, params.epsilon, params.batchSize, params.lambda, video);
 
-        String type = "Regression";
-        Integer epochs = 1000;
-        double lr = 0.01d;
-        String Model = "adam";
-        double beta1 = 0.99d;
-        double beta2 = 0.999d;
-        double epsilon = 1e-8d;
-        Integer batchSize = train;
-        double lambda = 1e-18d;
+        double[][] X = new double[params.train][params.feature];
+        double[][] Y = new double[params.train][params.output];
 
-        Model model = new Model(type, structure, functions, epochs, lr, Model, beta1,
-                beta2, epsilon, batchSize,
-                lambda, video);
+        double[][] x = new double[params.test][params.feature];
+        double[][] y = new double[params.test][params.output];
 
-        // Model model = new Model("C:\\Users\\horne\\Desktop\\Neural Networks\\wb.pt");
-
-        double[][] X = new double[train][feature];
-        double[][] Y = new double[train][output];
-
-        double[][] x = new double[test][feature];
-        double[][] y = new double[test][output];
-
-        for (int i = 0; i < train; i++) {
-            for (int j = 0; j < feature; j++) {
-                X[i][j] = rand.nextDouble(lower_bound, upper_bound);
+        for (int i = 0; i < params.train; i++) {
+            for (int j = 0; j < params.feature; j++) {
+                X[i][j] = rand.nextDouble(params.lower_bound, params.upper_bound);
             }
             Y[i] = function(X[i]);
         }
 
-        for (int i = 0; i < test; i++) {
-            for (int j = 0; j < feature; j++) {
-                x[i][j] = rand.nextDouble(lower_bound, upper_bound);
+        for (int i = 0; i < params.test; i++) {
+            for (int j = 0; j < params.feature; j++) {
+                x[i][j] = rand.nextDouble(params.lower_bound, params.upper_bound);
             }
             y[i] = function(x[i]);
         }
 
         Other.write(X, Y, "PYTHON/io.txt");
 
-        double[][] vX = new double[validate][feature];
-        double[][] vY = new double[validate][output];
-        double[][] predictedY = new double[validate][output];
+        double[][] vX = new double[params.validate][params.feature];
+        double[][] vY = new double[params.validate][params.output];
+        double[][] predictedY = new double[params.validate][params.output];
 
-        for (int i = 0; i < validate; i++) {
-            for (int j = 0; j < feature; j++) {
-                vX[i][j] = rand.nextDouble(lower_bound, upper_bound);
+        for (int i = 0; i < params.validate; i++) {
+            for (int j = 0; j < params.feature; j++) {
+                vX[i][j] = rand.nextDouble(params.lower_bound, params.upper_bound);
             }
             vY[i] = function(vX[i]);
         }
@@ -101,7 +85,7 @@ public class Main {
 
         Runnable plotTask = () -> {
             try {
-                Other.runPy(new String[] { "python", "PYTHON/plot.py", "PYTHON/io.txt", "Actual" });
+                Other.runPy(new String[]{"python", "PYTHON/plot.py", "PYTHON/io.txt", "Actual"});
             } catch (Exception e) {
                 System.err.println("Error in plot task: " + e);
                 e.printStackTrace();
@@ -113,11 +97,10 @@ public class Main {
             executorService.submit(plotTask);
 
         executorService.shutdown();
-
     }
 
     private static void process(Model model, double[][] X, double[][] Y, double[][] x, double[][] y, double[][] vX,
-            double[][] vY, double[][] predictedY, boolean video) {
+                                double[][] vY, double[][] predictedY, boolean video) {
 
         model.train(X, Y, x, y);
 
@@ -149,14 +132,70 @@ public class Main {
             Other.write(X, model.predict(X), "PYTHON/predicted.txt");
 
         if (X[0].length == 1 && Y[0].length == 1)
-            Other.runPy(new String[] { "python", "PYTHON/plot.py", "PYTHON/predicted.txt", "Predicted" });
+            Other.runPy(new String[]{"python", "PYTHON/plot.py", "PYTHON/predicted.txt", "Predicted"});
 
         if (video)
-            Other.runPy(new String[] { "python", "PYTHON/compile.py" });
+            Other.runPy(new String[]{"python", "PYTHON/compile.py"});
+    }
 
+    private static Parameters readParametersFromXML(String xmlFilePath) {
+        Parameters params = new Parameters();
+        try {
+            File xmlFile = new File(xmlFilePath);
+            DocumentBuilderFactory dbFactory = DocumentBuilderFactory.newInstance();
+            DocumentBuilder dBuilder = dbFactory.newDocumentBuilder();
+            Document doc = dBuilder.parse(xmlFile);
+
+            doc.getDocumentElement().normalize();
+
+            Element root = doc.getDocumentElement();
+
+            params.feature = Integer.parseInt(root.getElementsByTagName("feature").item(0).getTextContent());
+            params.output = Integer.parseInt(root.getElementsByTagName("output").item(0).getTextContent());
+            params.train = Integer.parseInt(root.getElementsByTagName("train").item(0).getTextContent());
+            params.test = Integer.parseInt(root.getElementsByTagName("test").item(0).getTextContent());
+            params.validate = Integer.parseInt(root.getElementsByTagName("validate").item(0).getTextContent());
+            params.lower_bound = Double.parseDouble(root.getElementsByTagName("lower_bound").item(0).getTextContent());
+            params.upper_bound = Double.parseDouble(root.getElementsByTagName("upper_bound").item(0).getTextContent());
+
+            params.structure = Arrays.stream(root.getElementsByTagName("structure").item(0).getTextContent().split(","))
+                    .map(Integer::parseInt).toArray(Integer[]::new);
+
+            params.functions = root.getElementsByTagName("functions").item(0).getTextContent().split(",");
+
+            params.type = root.getElementsByTagName("type").item(0).getTextContent();
+            params.epochs = Integer.parseInt(root.getElementsByTagName("epochs").item(0).getTextContent());
+            params.lr = Double.parseDouble(root.getElementsByTagName("lr").item(0).getTextContent());
+            params.Model = root.getElementsByTagName("Model").item(0).getTextContent();
+            params.beta1 = Double.parseDouble(root.getElementsByTagName("beta1").item(0).getTextContent());
+            params.beta2 = Double.parseDouble(root.getElementsByTagName("beta2").item(0).getTextContent());
+            params.epsilon = Double.parseDouble(root.getElementsByTagName("epsilon").item(0).getTextContent());
+            params.batchSize = Integer.parseInt(root.getElementsByTagName("batchSize").item(0).getTextContent());
+            params.lambda = Double.parseDouble(root.getElementsByTagName("lambda").item(0).getTextContent());
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return params;
+    }
+
+    private static class Parameters {
+        int feature;
+        int output;
+        int train;
+        int test;
+        int validate;
+        double lower_bound;
+        double upper_bound;
+        Integer[] structure;
+        String[] functions;
+        String type;
+        Integer epochs;
+        double lr;
+        String Model;
+        double beta1;
+        double beta2;
+        double epsilon;
+        Integer batchSize;
+        double lambda;
     }
 }
-
-
-
-
